+++ sdk/keyvault/Azure.Security.KeyVault.Keys/src/JsonWebKey.cs
@@ -37,6 +37,8 @@ namespace Azure.Security.KeyVault.Keys
         private const string DPropertyName = "d";
         private const string KPropertyName = "k";
         private const string TPropertyName = "key_hsm";
+        private const string AlgorithmPropertyName = "alg";
+        private const string PubPropertyName = "pub";
 
         private static readonly JsonEncodedText s_keyIdPropertyNameBytes = JsonEncodedText.Encode(KeyIdPropertyName);
         private static readonly JsonEncodedText s_keyTypePropertyNameBytes = JsonEncodedText.Encode(KeyTypePropertyName);
@@ -54,6 +56,8 @@ namespace Azure.Security.KeyVault.Keys
         private static readonly JsonEncodedText s_dPropertyNameBytes = JsonEncodedText.Encode(DPropertyName);
         private static readonly JsonEncodedText s_kPropertyNameBytes = JsonEncodedText.Encode(KPropertyName);
         private static readonly JsonEncodedText s_tPropertyNameBytes = JsonEncodedText.Encode(TPropertyName);
+        private static readonly JsonEncodedText s_algorithmPropertyNameBytes = JsonEncodedText.Encode(AlgorithmPropertyName);
+        private static readonly JsonEncodedText s_pubPropertyNameBytes = JsonEncodedText.Encode(PubPropertyName);
 
         private static readonly KeyOperation[] s_aesKeyOperation = { KeyOperation.Encrypt, KeyOperation.Decrypt, KeyOperation.WrapKey, KeyOperation.UnwrapKey };
         private static readonly KeyOperation[] s_rSAPublicKeyOperation = { KeyOperation.Encrypt, KeyOperation.Verify, KeyOperation.WrapKey };
@@ -253,6 +257,21 @@ namespace Azure.Security.KeyVault.Keys
         /// </summary>
         public byte[] T { get; set; }
 
+        #region AKP Key Parameters
+
+        /// <summary>
+        /// Gets or sets the algorithm identifier for an Algorithm Key Pair (AKP) key, such as ML-DSA.
+        /// This value is set when the <see cref="KeyType"/> is <see cref="KeyType.Akp"/> or <see cref="KeyType.AkpHsm"/>.
+        /// </summary>
+        public AkpAlgorithm? Algorithm { get; set; }
+
+        /// <summary>
+        /// Gets or sets the public key for an Algorithm Key Pair (AKP) key, such as ML-DSA.
+        /// </summary>
+        public byte[] Pub { get; set; }
+
+        #endregion
+
         internal bool HasPrivateKey
         {
             get
@@ -427,6 +446,12 @@ namespace Azure.Security.KeyVault.Keys
                     case TPropertyName:
                         T = Base64Url.Decode(prop.Value.GetString());
                         break;
+                    case AlgorithmPropertyName:
+                        Algorithm = prop.Value.GetString();
+                        break;
+                    case PubPropertyName:
+                        Pub = Base64Url.Decode(prop.Value.GetString());
+                        break;
                 }
             }
         }
@@ -501,6 +526,14 @@ namespace Azure.Security.KeyVault.Keys
             if (T != null)
             {
                 json.WriteString(s_tPropertyNameBytes, Base64Url.Encode(T));
+            }
+            if (Algorithm.HasValue)
+            {
+                json.WriteString(s_algorithmPropertyNameBytes, Algorithm.Value.ToString());
+            }
+            if (Pub != null)
+            {
+                json.WriteString(s_pubPropertyNameBytes, Base64Url.Encode(Pub));
             }
         }